pyproject.toml
[project]
requires-python=">=3.11"
name = "lang_agent" 
version = "0.1"   
dependencies = [
    "langchain>=1.0",
    "langchain_community",
    "langchain-openai",
    "langchain_mcp_adapters",
    "langchain-postgres",
    "agentbay",
    "httpx[socks]",
    "dashscope",
    "python-dotenv>=1.0.0",
    "websockets>=11.0.3",
    "mcp>=1.8.1",
    "pydantic<2.12",
    "mcp-proxy>=0.8.2",
    "loguru",
    "tyro",
    "faiss-cpu",
    "fastmcp",
    "fastapi",
    "matplotlib",
    "Pillow",
    "commentjson",
    "pandas",
    "asgiref",
    "psycopg[binary]",
    "psycopg-pool",
    "redis",
    "pika",
    "deepagents",
    "daytona",
    "langchain_daytona"
]

[tool.setuptools.packages.find]
include = ["lang_agent"]